+++ data/fanny.txt
@@ -1,3 +1,1 @@
-T | 0 | sleep
-D | 1 | eat |Dec 22, 2024 12:00
-E | 0 | dinner /from 2024-12-12 12:00 /to 2024-12-12 13:00|Dec 12, 2024 12:00|Dec 12, 2024 13:00
+E | 0 | eat |Feb 22, 2024 12:00|Feb 22, 2024 13:00
