+++ src/main/java/fanny/FannyException.java
@@ -1,3 +1,5 @@
+package fanny;
+
 public class FannyException extends Exception {
     public FannyException(String message) {
         super(message);
