+++ src/main/java/fanny/command/Command.java
@@ -1,3 +1,8 @@
+package fanny.command;
+
+import fanny.task.TaskList;
+import fanny.ui.Ui;
+
 import java.time.format.DateTimeFormatter;
 
 public abstract class Command {
