+++ src/main/java/fanny/command/DeadlineCommand.java
@@ -1,3 +1,10 @@
+package fanny.command;
+
+import fanny.ui.Ui;
+import fanny.task.TaskList;
+import fanny.task.Task;
+import fanny.task.Deadline;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
