+++ src/main/java/fanny/command/DeleteCommand.java
@@ -1,3 +1,8 @@
+package fanny.command;
+
+import fanny.task.TaskList;
+import fanny.ui.Ui;
+
 public class DeleteCommand extends Command {
     private String arguments;
