+++ src/main/java/fanny/task/Event.java
@@ -1,3 +1,5 @@
+package fanny.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
