+++ src/main/java/fanny/task/Task.java
@@ -1,3 +1,5 @@
+package fanny.task;
+
 public class Task {
 
     protected String description;
