+++ src/main/java/fanny/task/ToDo.java
@@ -1,3 +1,5 @@
+package fanny.task;
+
 public class ToDo extends Task {
 
     public ToDo(String description) {
